refact : 각각의 테이블 및 모델 타입을 단일 파일에서 관리
DESC
----
- 각각의 테이블 및 모델 타입을 단일 파일에서 관리

NOTE
----
- 각 필드의 unique함을 검증하기 위해 drizzle의 버전을 올려야 함
