--- packages/common/src/database/model/group.model.ts
@@ -1,6 +1,0 @@
-import { InferModel } from 'drizzle-orm';
-import { groups } from '../schema';
-
-type Group = InferModel<typeof groups>;
-
-export default Group;
