--- packages/common/src/database/model/index.ts
@@ -1,4 +1,0 @@
-import Group from './group.model';
-import Member from './member.model';
-import User from './user.model';
-export type { Group, Member, User };
