--- packages/common/src/database/model/member.model.ts
@@ -1,6 +1,0 @@
-import { InferModel } from 'drizzle-orm';
-import { groups } from '../schema';
-
-type Model = InferModel<typeof groups>;
-
-export default Model;
