--- packages/common/src/database/model/user.model.ts
@@ -1,6 +1,0 @@
-import { InferModel } from 'drizzle-orm';
-import { users } from '../schema';
-
-type User = InferModel<typeof users>;
-
-export default User;
