--- packages/common/src/database/schema/index.ts
@@ -1,4 +1,0 @@
-import groups from './group.schema';
-import members from './member.schema';
-import users from './user.schema';
-export { groups, members, users };
