--- packages/common/src/database/schema/user.schema.ts
@@ -1,8 +1,0 @@
-import { pgTable, serial, varchar } from 'drizzle-orm/pg-core';
-
-const users = pgTable('users', {
-  id: serial('id').primaryKey(),
-  email: varchar('email', { length: 320 }),
-});
-
-export default users;
